refactor: auto-split test_planning_collision.py
