+++ tests/unit/robotics/planning_collision/test_collision_result.py
@@ -1,0 +1,164 @@
+"""Tests for collision detection module.
+
+Tests cover:
+- Geometric primitives (Sphere, Box, Capsule, Cylinder, ConvexHull)
+- Distance computation between primitives
+- Collision checking
+- CollisionChecker configuration-space collision detection
+"""
+
+from __future__ import annotations
+
+import numpy as np
+import pytest
+from src.robotics.planning.collision import (
+    Box,
+    Capsule,
+    CollisionChecker,
+    CollisionCheckerConfig,
+    CollisionPair,
+    CollisionQuery,
+    CollisionQueryType,
+    CollisionResult,
+    ConvexHull,
+    Cylinder,
+    DistanceResult,
+    Sphere,
+    check_primitive_collision,
+    compute_primitive_distance,
+)
+
+# =============================================================================
+# Mock Engine for Testing
+# =============================================================================
+
+
+class MockCollisionEngine:
+    """Mock engine implementing CollisionCapable protocol."""
+
+    def __init__(self, n_dof: int = 7) -> None:
+        """Initialize mock engine."""
+        self._n_dof = n_dof
+        self._q = np.zeros(n_dof)
+        self._v = np.zeros(n_dof)
+        self._bodies: dict[str, dict] = {
+            "link1": {"position": np.array([0.0, 0.0, 0.5]), "radius": 0.1},
+            "link2": {"position": np.array([0.0, 0.0, 1.0]), "radius": 0.1},
+            "link3": {"position": np.array([0.0, 0.0, 1.5]), "radius": 0.1},
+        }
+
+    def get_state(self) -> tuple[np.ndarray, np.ndarray]:
+        """Get current state."""
+        return self._q.copy(), self._v.copy()
+
+    def set_state(self, q: np.ndarray, v: np.ndarray) -> None:
+        """Set state."""
+        self._q = q.copy()
+        self._v = v.copy()
+        # Update body positions based on q (simplified kinematics)
+        self._bodies["link1"]["position"] = np.array([q[0] * 0.1, q[1] * 0.1, 0.5])
+        self._bodies["link2"]["position"] = np.array([q[0] * 0.1, q[1] * 0.1, 1.0])
+        self._bodies["link3"]["position"] = np.array([q[2] * 0.1, q[3] * 0.1, 1.5])
+
+    def get_body_names(self) -> list[str]:
+        """Get body names."""
+        return list(self._bodies.keys())
+
+    def get_body_position(self, body_name: str) -> np.ndarray | None:
+        """Get body position."""
+        if body_name in self._bodies:
+            return self._bodies[body_name]["position"].copy()
+        return None
+
+    def get_body_collision_geometry(self, body_name: str) -> Sphere | None:
+        """Get collision geometry as sphere."""
+        if body_name in self._bodies:
+            return Sphere(
+                center=self._bodies[body_name]["position"],
+                radius=self._bodies[body_name]["radius"],
+            )
+        return None
+
+
+# =============================================================================
+# Sphere Tests
+# =============================================================================
+
+
+# =============================================================================
+# Box Tests
+# =============================================================================
+
+
+# =============================================================================
+# Capsule Tests
+# =============================================================================
+
+
+# =============================================================================
+# Cylinder Tests
+# =============================================================================
+
+
+# =============================================================================
+# ConvexHull Tests
+# =============================================================================
+
+
+# =============================================================================
+# Distance Computation Tests
+# =============================================================================
+
+
+# =============================================================================
+# CollisionPair Tests
+# =============================================================================
+
+
+# =============================================================================
+# CollisionResult Tests
+# =============================================================================
+
+
+class TestCollisionResult:
+    """Tests for CollisionResult dataclass."""
+
+    def test_create_collision_result(self) -> None:
+        """Test creating collision result."""
+        result = CollisionResult(
+            in_collision=True,
+            collision_pairs=[CollisionPair("link1", "link2")],
+            num_contacts=1,
+        )
+        assert result.in_collision
+        assert len(result.collision_pairs) == 1
+
+    @pytest.mark.parametrize(
+        "in_collision, pairs, match",
+        [
+            (True, [], "in_collision=True but no collision"),
+            (False, [CollisionPair("a", "b")], "in_collision=False but collision"),
+        ],
+        ids=["true-no-pairs", "false-with-pairs"],
+    )
+    def test_collision_result_inconsistent_state(
+        self, in_collision: bool, pairs: list, match: str
+    ) -> None:
+        """Test that inconsistent state raises error."""
+        with pytest.raises(ValueError, match=match):
+            CollisionResult(in_collision=in_collision, collision_pairs=pairs)
+
+
+# =============================================================================
+# DistanceResult Tests
+# =============================================================================
+
+
+# =============================================================================
+# CollisionQuery Tests
+# =============================================================================
+
+
+# =============================================================================
+# CollisionChecker Tests
+# =============================================================================
